ocr_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0